requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
json-repair>=0.3.0
flask>=2.0.0
pandas>=1.3.0
plotly>=5.3.0
spotipy>=2.23.0
